FIx https://github.com/SapienzaNLP/relik/issues/7
Fix https://github.com/SapienzaNLP/relik/issues/7

This was a fun one, had a missing len() which caused relation candidates to be 0 when candidates are trimmed due to max_length. Due to an older version of pytorch having issues with 0 dimensional tensors with einsum, it cause issues to some people on Colab.